caching much improved, and a much simpler structure than the old db format

+++ mutmut3.py
@@ -3,12 +3,14 @@ import gc
 import json
 import os
 from collections import defaultdict
+from dataclasses import dataclass
 from datetime import datetime
 from os import (
     makedirs,
     walk,
 )
 from pathlib import Path
+from typing import Dict
 
 from mutmut import (
     guess_paths_to_mutate,
@@ -33,17 +35,8 @@ def get_stats():
     return _stats
 
 
-def foo(a, *, b: int, **kwargs):
-    pass
-
-
-signature(foo)
-
-
-def create_mutants():
+def walk_files():
     paths = [guess_paths_to_mutate()]
-    next_id = 0
-    function_by_id = {}
     for path in paths:
         for root, dirs, files in walk(path):
             for filename in files:
@@ -53,8 +46,14 @@ def create_mutants():
                     continue
                 if filename.startswith('test_.py'):
                     continue
-                next_id = create_mutants_for_file(Path(root) / filename, next_id=next_id, function_by_id=function_by_id)
-    return next_id, function_by_id
+                yield Path(root) / filename
+
+
+def create_mutants():
+    function_by_id = {}
+    for path in walk_files():
+        create_mutants_for_file(path, function_by_id=function_by_id)
+    return function_by_id
 
 
 def write_trampoline(out, orig_name, mutant_names):
@@ -75,11 +74,11 @@ def {orig_name}(*args, **kwargs):
 
 
 def write_mutant(out, c, mutation_id, next_id, mutant_names, orig_name):
-    if not mutation_id.subject_stack:
+    if not mutation_id.subject:
         return
     c.mutation_id = mutation_id
     new_code, number = mutate(c)
-    node = mutation_id.subject_stack[-1]
+    node = mutation_id.subject
     node.name.value += f'_mutant_{next_id}'
     mutant_names.append(node.name.value)
     # assert number == 1, number
@@ -93,8 +92,8 @@ def write_mutant(out, c, mutation_id, next_id, mutant_names, orig_name):
     return node.get_code()
 
 
-def write_original_alias(out, last_subject_stack):
-    orig_name = last_subject_stack[-1].name.value
+def write_original_alias(out, last_subject):
+    orig_name = last_subject.name.value
     print(f'{orig_name}_orig = {orig_name}', file=out)  # the trampoline will then overwrite the original
 
 
@@ -112,25 +111,27 @@ def trampoline(orig, mutants, *args, **kwargs):
     elif mutant_under_test == 'stats':
         from __main__ import record_trampoline_hit
         record_trampoline_hit(orig.__module__ + '.' + orig.__name__)
-        mutant_under_test = -1
-    return mutants.get(int(mutant_under_test), orig)(*args, **kwargs)
+        mutant_under_test = '..matches nothing..'
+    prefix = orig.__module__ + '.' + orig.__name__ + '$'
+    if not mutant_under_test.startswith(prefix):
+        return orig(*args, **kwargs)
+    mutant_id = mutant_under_test[len(prefix):]
+    return mutants[int(mutant_id)](*args, **kwargs)
 
 """, file=out)
 
 
-def create_mutants_for_file(filename, next_id, function_by_id):
+def create_mutants_for_file(filename, function_by_id):
     output_path = Path('mutants') / filename
     makedirs(output_path.parent, exist_ok=True)
 
     input_stat = os.stat(filename)
 
-    meta_filename = str(output_path) + '.next_id'
+    if output_path.exists() and output_path.stat().st_mtime == input_stat.st_mtime:
+        # print('    skipped', output_path, 'already up to date')
+        return
 
-    # TODO: cache
-    # if output_path.exists() and output_path.stat().st_mtime == input_stat.st_mtime:
-    #     print('    skipped', output_path, 'already up to date')
-    #     with open(meta_filename) as f:
-    #         return int(f.read().strip())
+    num_mutants_by_function = {}
 
     with open(output_path, 'w') as out:
         c = Context(filename=filename)
@@ -141,39 +142,43 @@ def create_mutants_for_file(filename, next_id, function_by_id):
         mutation_ids = list_mutations(c)
 
         mutant_names = []
-        last_subject_stack = None
+        last_subject = None
 
         last_code = None
 
+        next_id = 0
+
         for mutation_id in mutation_ids:
-            if not mutation_id.subject_stack:
+            if not mutation_id.subject:
                 continue
 
             # TODO: mutate methods too!, then we have a classdef then a funcdef in the stack
-            if mutation_id.subject_stack[0].type != 'funcdef':
+            if mutation_id.subject.type != 'funcdef':
                 continue
 
-            if last_subject_stack != mutation_id.subject_stack:
-                if last_subject_stack:
-                    write_original_alias(out, last_subject_stack)
+            if last_subject != mutation_id.subject:
+                if last_subject:
+                    write_original_alias(out, last_subject)
 
                 if mutant_names:
                     write_trampoline(out, orig_name, mutant_names)
 
-                orig_name = mutation_id.subject_stack[0].name.value
+                orig_name = mutation_id.subject.name.value
                 mutant_names = []
-                last_subject_stack = mutation_id.subject_stack
+                last_subject = mutation_id.subject
+                next_id = 0
 
             code = write_mutant(out, c, mutation_id, next_id, mutant_names, orig_name)
             module_name = str(filename)[:-len(filename.suffix)].replace(os.sep, ".").replace('.__init__', '')
-            function_by_id[next_id] = f'{module_name}.{orig_name}'
             assert last_code != code
             last_code = code
 
+            num_mutants_by_function[f'{module_name}.{orig_name}'] = next_id
+
             next_id += 1
 
-        if last_subject_stack:
-            write_original_alias(out, last_subject_stack)
+        if last_subject:
+            write_original_alias(out, last_subject)
 
         if mutant_names:
             write_trampoline(out, orig_name, mutant_names)
@@ -186,11 +191,41 @@ def create_mutants_for_file(filename, next_id, function_by_id):
             print(output_path, 'has invalid syntax: ', e)
             exit(1)
 
-    os.utime(output_path, (input_stat.st_atime, input_stat.st_mtime))
+    meta_filename = str(output_path) + '.meta'
     with open(meta_filename, 'w') as f:
-        f.write(str(next_id))
+        json.dump(dict(
+            num_mutants_by_function=num_mutants_by_function,
+            result_by_key={},
+        ), f)
 
-    return next_id
+    os.utime(output_path, (input_stat.st_atime, input_stat.st_mtime))
+
+
+class MutationData:
+    def __init__(self, *, path):
+        self.meta_path = Path('mutants') / (str(path) + '.meta')
+        with open(self.meta_path) as f:
+            self.meta = json.load(f)
+
+        self.key_by_pid = {}
+        self.result_by_key = self.meta.pop('result_by_key')
+        self.num_mutants_by_function = self.meta.pop('num_mutants_by_function')
+        assert not self.meta, self.meta  # We should read all the data!
+
+    def register_pid(self, *, pid, key):
+        self.key_by_pid[pid] = key
+
+    def register_result(self, *, pid, exit_code):
+        self.result_by_key[self.key_by_pid[pid]] = (0xFF00 & exit_code) >> 8  # The high byte contains the exit code
+        self.save()
+
+
+    def save(self):
+        with open(self.meta_path, 'w') as f:
+            json.dump(dict(
+                result_by_key=self.result_by_key,
+                num_mutants_by_function=self.num_mutants_by_function,
+            ), f)
 
 
 def mutmut_3():
@@ -198,8 +233,7 @@ def mutmut_3():
 
     start = datetime.now()
     print('generating mutants...')
-    # TODO: read from db!!!!
-    next_id, function_by_id = create_mutants()
+    create_mutants()
     time = datetime.now() - start
     print('mutation generation', time)
 
@@ -209,18 +243,23 @@ def mutmut_3():
 
     sys.path.insert(0, os.path.abspath('mutants'))
 
+    # TODO: run these three steps only if we have mutants to test
     print('running baseline...')
-    os.environ['MUTANT_UNDER_TEST'] = '-1'
-    if hammett.main(quiet=True, fail_fast=True, disable_assert_analyze=True) != 0:
+    os.environ['MUTANT_UNDER_TEST'] = '..match nothing..'
+    if hammett.main(quiet=True, fail_fast=True, disable_assert_analyze=True, use_cache=False) != 0:
         print("FAILED")
         return
     print('done')
 
     print('running forced fail test')
     os.environ['MUTANT_UNDER_TEST'] = 'fail'
-    if hammett.main(fail_fast=True, disable_assert_analyze=True) == 0:
-        print("FAILED")
-        return
+    try:
+        if hammett.main(quiet=True, fail_fast=True, disable_assert_analyze=True, use_cache=False) == 0:
+            print("FAILED")
+            return
+    except Exception as e:
+        # We get here if there's a mutant in the setup code path
+        assert e.args[0] == 'Failed programmatically'
     print('done')
 
     print('running stats...')
@@ -233,22 +272,14 @@ def mutmut_3():
             tests_by_function[function].add(_name)
         _stats.clear()
 
-    if hammett.main(quiet=True, fail_fast=True, disable_assert_analyze=True, post_test_callback=post_test_callback) != 0:
+    if hammett.main(quiet=True, fail_fast=True, disable_assert_analyze=True, post_test_callback=post_test_callback, use_cache=False) != 0:
         print("FAILED")
         return
     print('done')
 
-    db_path = Path('mutants') / 'db.json'
-    try:
-        with open(db_path) as f:
-            db = json.loads(f.read())
-            result_by_key = {int(k): v for k, v in db['result_by_key'].items()}
-    except FileNotFoundError:
-        result_by_key = {}
-
     def read_one_child_exit_status():
-        pid, status = os.wait()
-        result_by_key[key_from_pid[pid]] = (0xFF00 & status) >> 8  # The high byte contains the exit code
+        pid, exit_code = os.wait()
+        mutation_data_by_pid[pid].register_result(pid=pid, exit_code=exit_code)
 
     hammett_kwargs = hammett.main_setup(
         quiet=True,
@@ -257,81 +288,76 @@ def mutmut_3():
         use_cache=False,
     )
 
-    key_from_pid = {}
+    mutation_data_by_path : Dict[str, MutationData] = {}
+    mutation_data_by_pid : Dict[int, MutationData] = {}  # many pids map to one MutationData
     running_children = 0
-    max_children = 8
+    max_children = os.cpu_count()
 
     start = datetime.now()
 
     gc.freeze()
 
-    count = 0
+    total_count = 0
 
     try:
         print('Running mutation testing...')
-        for key in tqdm(range(next_id)):
-            if key in result_by_key:
-                continue
+        for path in tqdm(walk_files()):  # TODO: now the progress bar is per file, which sucks a bit
+            assert path not in mutation_data_by_path
+            m = MutationData(path=path)
+            mutation_data_by_path[str(path)] = m
 
-            pid = os.fork()
-            if not pid:
-                sys.path.insert(0, os.path.abspath('mutants'))
-                # In the child
-                os.environ['MUTANT_UNDER_TEST'] = str(key)
+            for function, count in m.num_mutants_by_function.items():
+                for i in range(count):
+                    key = f'{function}${i}'
 
-                # TODO: this is needed for non-memory DBs
-                hammett.Config.workerinput = dict(workerinput=f'_{key}')
+                    if key in m.result_by_key:
+                        continue
 
-                result = hammett.main_run_tests(**hammett_kwargs, tests=tests_by_function[function_by_id[key]])
-                if result != 0:
-                    # TODO: write failure information to stdout?
-                    pass
-                os._exit(result)
-            else:
-                key_from_pid[pid] = key
-                running_children += 1
+                    pid = os.fork()
+                    if not pid:
+                        sys.path.insert(0, os.path.abspath('mutants'))
+                        # In the child
+                        os.environ['MUTANT_UNDER_TEST'] = key
 
-            if running_children >= max_children:
-                read_one_child_exit_status()
-                count += 1
-                running_children -= 1
+                        hammett.Config.workerinput = dict(workerinput=f'_{key}')
+
+                        result = hammett.main_run_tests(**hammett_kwargs, tests=tests_by_function[function])
+                        if result != 0:
+                            # TODO: write failure information to stdout?
+                            pass
+                        os._exit(result)
+                    else:
+                        mutation_data_by_pid[pid] = m
+                        m.register_pid(pid=pid, key=key)
+                        running_children += 1
+
+                    if running_children >= max_children:
+                        read_one_child_exit_status()
+                        total_count += 1
+                        running_children -= 1
 
         try:
             while running_children:
                 read_one_child_exit_status()
-                count += 1
+                total_count += 1
                 running_children -= 1
         except ChildProcessError:
             pass
     except KeyboardInterrupt:
         print('aborting...')
-    finally:
-        with open(db_path, 'w') as f:
-            json.dump(
-                dict(
-                    version=1,
-                    result_by_key=result_by_key,
-                    next_id=next_id,
-                    function_by_id=function_by_id,
-                ),
-                f
-            )
 
     t = datetime.now() - start
 
-    covered = {k for k, v in result_by_key.items() if v != 0}
-    not_covered = {k for k, v in result_by_key.items() if v == 0}
-    print('covered:', covered)
-    print('not covered:', not_covered)
+    covered = {k for m in mutation_data_by_path.values() for k, v in m.result_by_key.items() if v != 0}
+    not_covered = {k for m in mutation_data_by_path.values() for k, v in m.result_by_key.items() if v == 0}
 
     print('number of covered:', len(covered))
     print('number of not covered:', len(not_covered))
     print('%:', len(not_covered) / (len(covered) + len(not_covered)) * 100)
 
     print('time:', t)
-    print('next ID:', next_id)
-    print('number of tested mutants:', count)
-    print('mutations/s:', count / t.total_seconds())
+    print('number of tested mutants:', total_count)
+    print('mutations/s:', total_count / t.total_seconds())
 
 
 if __name__ == '__main__':
